Move prompt to new file
